# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FITS Backend API
  version: 1.0.0
  description: |
    # FITS (Flexible IT Training System) Backend API

    Production-ready REST API for managing students, teachers, authentication, and signing workflows.

    ## Features
    - JWT-based authentication with refresh tokens
    - Role-based access control (Admin, Teacher, Student)
    - Invitation system for user onboarding
    - Student and teacher management with pagination
    - Digital signing workflow for reports
    - Rate limiting and security headers

    ## Authentication Flow
    1. **Bootstrap Admin**: `POST /api/v1/bootstrap/init` (one-time setup)
    2. **Login**: `POST /api/v1/auth/login` (get access + refresh tokens)
    3. **Use Access Token**: Include `Authorization: Bearer {access_token}` header
    4. **Refresh Token**: `POST /api/v1/auth/refresh` when access token expires
    5. **Logout**: `POST /api/v1/auth/logout` to invalidate tokens

    ## Rate Limiting
    - Global: 100 requests/minute per IP
    - Per-user: Varies by role (Admin: 200/min, Teacher: 100/min, Student: 50/min)

    ## Pagination
    All list endpoints support:
    - `page` (default: 1, min: 1)
    - `limit` (default: 20, max: 100)
  termsOfService: https://fits.example.com/terms
  contact:
    name: API Support Team
    email: support@fits.example.com
    url: https://fits.example.com/support
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/bootstrap/init:
    post:
      summary: Initialize admin certificate
      description: |
        Generate admin RSA keypair and return admin token.
        **Can only be called once** - subsequent calls will fail.

        This is the first endpoint to call when setting up the system.
        Save the returned `admin_token` securely.
      operationId: bootstrapInit
      tags: [Bootstrap]
      security: []
      responses:
        '200':
          description: Admin certificate generated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BootstrapResponse'
              examples:
                success:
                  summary: Successful bootstrap
                  value:
                    success: true
                    message: Admin certificate generated successfully
                    data:
                      admin_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiYWRtaW4tMTIzNCIsInJvbGUiOiJhZG1pbiIsInRva2VuX3R5cGUiOiJhY2Nlc3MiLCJleHAiOjE3Mjk2ODk2MDAsImlhdCI6MTcyOTY4NjAwMH0.abc123xyz
                      message: Admin certificate generated successfully
                      public_key_path: ./configs/keys/admin.pub
        '400':
          description: Already initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                alreadyInitialized:
                  summary: System already initialized
                  value:
                    success: false
                    code: 400
                    error: Bad Request
                    details: Admin already initialized. Bootstrap can only be called once.
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== AUTHENTICATION ====================
  /api/v1/auth/login:
    post:
      summary: User login
      description: |
        Authenticate with username and password.
        Returns access token (short-lived) and refresh token (long-lived).
      operationId: authLogin
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              studentLogin:
                summary: Student login
                value:
                  username: max.mustermann
                  password: SecurePassword123!
              teacherLogin:
                summary: Teacher login
                value:
                  username: anna.schmidt
                  password: TeacherPass456!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
              examples:
                studentSuccess:
                  summary: Student login successful
                  value:
                    success: true
                    message: Login successful
                    data:
                      access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                      refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                      token_type: Bearer
                      expires_in: 3600
                      role: student
                      user_id: 550e8400-e29b-41d4-a716-446655440000
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidCredentials:
                  summary: Wrong username or password
                  value:
                    success: false
                    code: 401
                    error: Unauthorized
                    details: Invalid credentials
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/auth/refresh:
    post:
      summary: Refresh access token
      description: |
        Get a new access token using a valid refresh token.
        The refresh token remains valid and is returned in the response.
      operationId: authRefresh
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            examples:
              refreshExample:
                summary: Refresh token request
                value:
                  refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
              examples:
                success:
                  summary: New access token generated
                  value:
                    success: true
                    message: Token refreshed successfully
                    data:
                      access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                      refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                      token_type: Bearer
                      expires_in: 3600
                      role: student
                      user_id: 550e8400-e29b-41d4-a716-446655440000
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                expired:
                  summary: Refresh token expired
                  value:
                    success: false
                    code: 401
                    error: Unauthorized
                    details: Refresh token expired
                invalid:
                  summary: Invalid refresh token
                  value:
                    success: false
                    code: 401
                    error: Unauthorized
                    details: Invalid refresh token
  /api/v1/auth/logout:
    post:
      summary: User logout
      description: |
        Invalidate all refresh tokens for the authenticated user.
        Access tokens remain valid until expiry (cannot be revoked).
      operationId: authLogout
      tags: [Authentication]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              examples:
                success:
                  summary: Successfully logged out
                  value:
                    success: true
                    message: Logged out successfully
                    data: null
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== INVITATIONS ====================
  /api/v1/admin/invite:
    post:
      summary: Create invitation
      description: |
        Create invitation link for new student or teacher (Admin only).

        **Required fields by role:**
        - **Student**: email, first_name, last_name, role=student, teacher_uuid
        - **Teacher**: email, first_name, last_name, role=teacher, department
      operationId: createInvitation
      tags: [Invitations]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvitationRequest'
            examples:
              studentInvite:
                summary: Invite a student
                value:
                  email: max.mustermann@example.com
                  first_name: Max
                  last_name: Mustermann
                  role: student
                  teacher_uuid: 550e8400-e29b-41d4-a716-446655440010
              teacherInvite:
                summary: Invite a teacher
                value:
                  email: anna.schmidt@example.com
                  first_name: Anna
                  last_name: Schmidt
                  role: teacher
                  department: Computer Science
      responses:
        '201':
          description: Invitation created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CreateInvitationResponse'
              examples:
                success:
                  summary: Invitation created
                  value:
                    success: true
                    message: Invitation created successfully
                    data:
                      invitation_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                      invitation_link: https://fits.example.com/invite/eyJhbGc...
                      expires_at: "2025-10-25T12:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: Validation error - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingTeacherUuid:
                  summary: Student missing teacher_uuid
                  value:
                    success: false
                    code: 422
                    error: Validation Error
                    details: "teacher_uuid is required for student invitations"
                missingDepartment:
                  summary: Teacher missing department
                  value:
                    success: false
                    code: 422
                    error: Validation Error
                    details: "department is required for teacher invitations"
  /api/v1/invite/{token}:
    get:
      summary: Get invitation details
      description: |
        Retrieve invitation information by token.
        Used by frontend to show invitation details before registration.
      operationId: getInvitation
      tags: [Invitations]
      security: []
      parameters:
        - name: token
          in: path
          required: true
          description: JWT invitation token
          schema:
            type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Invitation details retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/InvitationResponse'
              examples:
                studentInvite:
                  summary: Student invitation
                  value:
                    success: true
                    data:
                      email: max.mustermann@example.com
                      first_name: Max
                      last_name: Mustermann
                      role: student
                      teacher_uuid: 550e8400-e29b-41d4-a716-446655440010
                      expires_at: "2025-10-25T12:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                expired:
                  value:
                    success: false
                    code: 401
                    error: Unauthorized
                    details: Invitation token expired
        '404':
          description: Invitation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  value:
                    success: false
                    code: 404
                    error: Not Found
                    details: Invitation not found or already used
  /api/v1/invite/{token}/complete:
    post:
      summary: Complete invitation
      description: |
        Complete user registration with username and password.
        Creates user account and invalidates invitation token.
      operationId: completeInvitation
      tags: [Invitations]
      security: []
      parameters:
        - name: token
          in: path
          required: true
          description: JWT invitation token
          schema:
            type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteInvitationRequest'
            examples:
              registration:
                summary: Complete registration
                value:
                  username: max.mustermann
                  password: SecurePassword123!
      responses:
        '200':
          description: Registration completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              examples:
                success:
                  value:
                    success: true
                    message: Registration completed successfully
                    data: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                conflict:
                  value:
                    success: false
                    code: 409
                    error: Conflict
                    details: Username already exists
        '422':
          $ref: '#/components/responses/ValidationError'

  # ==================== STUDENTS ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication.

        **Format**: `Authorization: Bearer {access_token}`

        **How to get tokens:**
        1. Login: `POST /api/v1/auth/login`
        2. Use returned `access_token` in Authorization header
        3. Refresh when expired: `POST /api/v1/auth/refresh`
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidJson:
              summary: Invalid JSON
              value:
                success: false
                code: 400
                error: Bad Request
                details: Invalid JSON format
            invalidUuid:
              summary: Invalid UUID format
              value:
                success: false
                code: 400
                error: Bad Request
                details: Invalid UUID format
    Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingToken:
              summary: Missing authorization header
              value:
                success: false
                code: 401
                error: Unauthorized
                details: Missing authorization token
            invalidToken:
              summary: Invalid or expired token
              value:
                success: false
                code: 401
                error: Unauthorized
                details: Invalid or expired token
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            insufficientPermissions:
              summary: Admin role required
              value:
                success: false
                code: 403
                error: Forbidden
                details: Admin role required
...
    ValidationError:
      description: Validation error - invalid field values
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            passwordTooShort:
              summary: Password validation failed
              value:
                success: false
                code: 422
                error: Validation Error
                details: Password must be at least 8 characters long
            invalidEmail:
              summary: Email validation failed
              value:
                success: false
                code: 422
                error: Validation Error
                details: Invalid email format
            missingField:
              summary: Required field missing
              value:
                success: false
                code: 422
                error: Validation Error
                details: "first_name is required"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            serverError:
              summary: Generic server error
              value:
                success: false
                code: 500
                error: Internal Server Error
                details: An unexpected error occurred. Please try again later.
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
        data:
          type: object
          nullable: true
          description: Response payload (varies by endpoint)
    ErrorResponse:
      type: object
      required: [success, code, error]
      properties:
        success:
          type: boolean
          example: false
        code:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type/category
          example: Bad Request
        details:
          type: string
          description: Human-readable error details
          example: Invalid email format
...
    BootstrapResponse:
      type: object
      required: [admin_token, message, public_key_path]
      properties:
        admin_token:
          type: string
          description: JWT access token for admin user
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        message:
          type: string
          example: Admin certificate generated successfully
        public_key_path:
          type: string
          description: Path to generated RSA public key
          example: ./configs/keys/admin.pub

    # ==================== AUTHENTICATION ====================
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
          description: User's username
          example: max.mustermann
        password:
          type: string
          format: password
          minLength: 8
          description: User's password (min 8 characters)
          example: SecurePassword123!
    LoginResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in, role, user_id]
      properties:
        access_token:
          type: string
          description: JWT access token (short-lived)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          description: JWT refresh token (long-lived)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          enum: [Bearer]
          example: Bearer
        expires_in:
          type: integer
          description: Access token expiry in seconds
          example: 3600
        role:
          type: string
          enum: [admin, teacher, student]
          description: User's role
          example: student
        user_id:
          type: string
          format: uuid
          description: User's unique identifier
          example: 550e8400-e29b-41d4-a716-446655440000
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          description: Valid refresh token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...

    # ==================== INVITATIONS ====================
    CreateInvitationRequest:
      type: object
      required: [email, first_name, last_name, role]
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: max.mustermann@example.com
        first_name:
          type: string
          minLength: 1
          maxLength: 100
          description: User's first name
          example: Max
        last_name:
          type: string
          minLength: 1
          maxLength: 100
          description: User's last name
          example: Mustermann
        role:
          type: string
          enum: [student, teacher]
          description: User's role
          example: student
        teacher_uuid:
          type: string
          format: uuid
          description: Required for student invitations - UUID of assigned teacher
          example: 550e8400-e29b-41d4-a716-446655440010
        department:
          type: string
          minLength: 1
          maxLength: 100
          description: Required for teacher invitations - teacher's department
          example: Computer Science
    CreateInvitationResponse:
      type: object
      required: [invitation_token, invitation_link, expires_at]
      properties:
        invitation_token:
          type: string
          description: JWT invitation token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        invitation_link:
          type: string
          format: uri
          description: Complete invitation URL to send to user
          example: https://fits.example.com/invite/eyJhbGc...
        expires_at:
          type: string
          format: date-time
          description: Token expiration timestamp
          example: "2025-10-25T12:00:00Z"
    InvitationResponse:
      type: object
      required: [email, first_name, last_name, role, expires_at]
      properties:
        email:
          type: string
          format: email
          example: max.mustermann@example.com
        first_name:
          type: string
          example: Max
        last_name:
          type: string
          example: Mustermann
        role:
          type: string
          enum: [student, teacher]
          example: student
        teacher_uuid:
          type: string
          format: uuid
          nullable: true
          description: Present for student invitations
          example: 550e8400-e29b-41d4-a716-446655440010
        department:
          type: string
          nullable: true
          description: Present for teacher invitations
          example: Computer Science
        expires_at:
          type: string
          format: date-time
          example: "2025-10-25T12:00:00Z"
    CompleteInvitationRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
          description: Desired username (must be unique)
          example: max.mustermann
        password:
          type: string
          format: password
          minLength: 8
          description: Password (min 8 characters, must include letters and numbers)
          example: SecurePassword123!

    # ==================== STUDENTS ====================
...